requirements.txt
pandas
sqlalchemy
pymysql
langchain
langchain-openai
faiss-cpu
python-dotenv
pyyaml
sqlglot